+++ dist-newstyle/packagedb/ghc-9.4.8/haskell-demo-0.1.0.0-inplace.conf
@@ -1,0 +1,38 @@
+name:                 haskell-demo
+version:              0.1.0.0
+visibility:           public
+id:                   haskell-demo-0.1.0.0-inplace
+key:                  haskell-demo-0.1.0.0-inplace
+license:              MIT
+maintainer:           mitomi.yuuto2003@gmail.com
+author:               Yuto729
+abi:                  inplace
+exposed:              True
+exposed-modules:      Poker Utils
+import-dirs:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\build
+
+library-dirs:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\build
+    C:\ghcup\msys64\mingw64\lib
+
+library-dirs-static:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\build
+    C:\ghcup\msys64\mingw64\lib
+
+dynamic-library-dirs:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\build
+    C:\ghcup\msys64\mingw64\lib
+
+data-dir:             C:\Users\mitom\haskell_demo\.
+hs-libraries:         HShaskell-demo-0.1.0.0-inplace
+include-dirs:         C:\ghcup\msys64\mingw64\include
+depends:
+    base-4.17.2.1
+    random-1.2.1.2-dda1dec15cd036d8dff9d229ba9f17ab5b775a8d
+
+haddock-interfaces:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\doc\html\haskell-demo\haskell-demo.haddock
+
+haddock-html:
+    C:\Users\mitom\haskell_demo\dist-newstyle\build\x86_64-windows\ghc-9.4.8\haskell-demo-0.1.0.0\doc\html\haskell-demo
