+++ haskell-demo.cabal
@@ -53,6 +53,14 @@ extra-doc-files:    CHANGELOG.md
 common warnings
     ghc-options: -Wall
 
+library
+    exposed-modules:    Poker
+                       ,Utils
+    build-depends:      base ^>=4.17.2.1
+                       ,random
+    hs-source-dirs:     src
+    default-language:   Haskell2010
+
 executable haskell-demo
     -- Import common warning flags.
     import:           warnings
@@ -60,15 +68,12 @@ executable haskell-demo
     -- .hs or .lhs file containing the Main module.
     main-is:          Main.hs
 
-    -- Modules included in this executable, other than Main.
-    -- other-modules:
-
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
 
     -- Other library packages from which modules are imported.
-    build-depends:    base ^>=4.17.2.1
-                     ,random
+    build-depends:   base ^>=4.17.2.1,
+                     haskell-demo
 
     -- Directories containing source files.
     hs-source-dirs:   app
